requirements.txt
streamlit
easyocr
pyttsx3
opencv-python
opencv-contrib-python
Pillow
numpy
transformers
torch
torchvision
deep-translator
requests
gtts
reportlab
python-docx
wordcloud
plotly
pandas
